requirements.txt
langchain-openai
chromadb
python-dotenv
tiktoken
langchain
langchain_community